Add files via upload
First, randomforest.py code needs to be implemented. The input directory for this file is Training_WFDB. It also calls get_12ECG_features. 
The command to run this code on terminal is: python3 randomforest.py ./Training_WFDB
Second, driver.py needs to be run. The input directory for this file is also Training_WFDB and the output directory needs to be specified (it generates a directory based on the specified name). This program calls get_12ECG_features and run_12ECG_classifier.py. 
The command to run this code on terminal is: python3 driver.py ./Training_WFDB ./output